Add a rust implementation.
Uses RocksDB instead so the two versions can't use the same database.

The interface is slightly different for the persistent version but
probably fine in practice. May reintroduce soft_remove/load later, but
the idea this time was to make them as compatible as possible.

Name will be changed eventually.